4.0.0:制作牧师施法界面充能显示效果

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/spells/AuraOfProtection.java
@@ -48,12 +48,12 @@ public class AuraOfProtection extends ClericSpell {
 	public String desc() {
 		int dmgReduction = 10 + 10*Dungeon.hero.pointsInTalent(Talent.AURA_OF_PROTECTION);
 		int glyphPow = 25 + 25*Dungeon.hero.pointsInTalent(Talent.AURA_OF_PROTECTION);
-		return Messages.get(this, "desc", dmgReduction, glyphPow) + "\n\n" + Messages.get(this, "charge_cost", (int)chargeUse(Dungeon.hero));
+		return Messages.get(this, "desc", dmgReduction, glyphPow) + "\n\n" + Messages.get(this, "charge_cost", chargeUse(Dungeon.hero));
 	}
 
 	@Override
-	public float chargeUse(Hero hero) {
-		return 2f;
+	public int chargeUse(Hero hero) {
+		return 2;
 	}
 
 	@Override
