+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/spells/Smite.java
@@ -40,7 +40,6 @@ import com.shatteredpixel.shatteredpixeldungeon.ui.HeroIcon;
 import com.shatteredpixel.shatteredpixeldungeon.utils.GLog;
 import com.watabou.noosa.audio.Sample;
 import com.watabou.utils.Callback;
-import com.watabou.utils.Random;
 
 public class Smite extends TargetedClericSpell {
 
@@ -72,13 +71,13 @@ public class Smite extends TargetedClericSpell {
 		} else {
 			desc = Messages.get(this, "desc",2, min, max);//+0
 		}
-		desc += "\n\n" + Messages.get(this, "charge_cost", (int)chargeUse(Dungeon.hero));
+		desc += "\n\n" + Messages.get(this, "charge_cost", chargeUse(Dungeon.hero));
 		return desc;
 	}
 
 	@Override
-	public float chargeUse(Hero hero) {
-		return 2f;
+	public int chargeUse(Hero hero) {
+		return 2;
 	}
 
 	@Override
